Moved all relevant functions from CSP to SolarTrough.py renamed: - SolarTrough to solar_trough - SolarTower to solar_tower updated imports: - CombiCSP.ipynb - CSP50Compare.py - CSPCret_np.py
Removed commented out imports from CSPecon.py
