+++ CombiCSP/__init__.py
@@ -1,1 +1,3 @@
 from CombiCSP.misc import CtoK, OutputContainer
+from CombiCSP.solar_tower import SolarTowerCalcs
+from CombiCSP.solar_trough import SolarTroughCalcs
