+++ CombiCSP/solar_tower.py
