+++ examples/CSPecon.py
@@ -2,11 +2,8 @@
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd 
-# from CombiCSP.SolarGeometry import *
-# import CombiCSP.SolarGeometry as sgh
-# import CombiCSP.CSP as cspC
+
 from CombiCSP.storage import pcm_rho
-# from CSPCret import *
 from CSPCret import Ptrough, Etrough
